feat: 아이템 상세조회 API 연동 (#59)

+++ app/src/main/java/com/hyeeyoung/wishboard/presentation/cart/CartScreen.kt
@@ -181,7 +181,7 @@ fun ItemCountController(count: Int, onChangeItemCount: (Int) -> Unit) {
 }
 
 @Composable
-fun CartTotalDisplay(totalCount: Int, totalPrice: Int) {
+fun CartTotalDisplay(totalCount: Int, totalPrice: Long) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
